+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1
 
+* Up supported PHP version to >= 7.4
 * Add `FUL` order type for EBICS 2.5
 * Add methods `YCT`, `ZSR`, `Z54` order type for EBICS 3.0
 * Major update for keyring manager. Added Array & File keyring managers instead of keyring.
@@ -7,3 +8,4 @@
 * Add $storageCallback for download methods that handle acknowledge.
 * Fix UTF-8 encoding.
 * Added CurlHttpClient and PsrHttpClient to use standard client.
+* Updated AbstractX509Generator to handle custom options.